requirements.txt
django
djangorestframework
spacy
pdfplumber
